@@ -24,10 +24,10 @@
     "date-fns": "^2.30.0",
     "graphql": "^16.8.1",
     "js-cookie": "^3.0.5",
-    "next": "14.0.0",
+    "next": "^14.0.1",
     "next-auth": "^4.24.4",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.47.0",
     "swr": "^2.2.4"
   },
