@@ -1,24 +1,25 @@
+'use client';
+
 import NextLink from 'next/link';
-import { useRouter } from 'next/router';
-import { useContext, useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import { AuthLayout } from '@/components/layouts/AuthLayout';
-import { isEmail, isPassword } from '@/utils/validations';
 import { ErrorOutline } from '@mui/icons-material';
 import { Box, Button, Chip, Grid, TextField, Typography } from '@mui/material';
-
+import { useContext, useState } from 'react';
 import { AuthContext } from '../../context';
+import { AuthLayout } from '../../components/layouts/AuthLayout';
+import { isEmail, isPassword } from '../../utils/validations';
+import { useRouter } from 'next/navigation';
+
 
 type FormData = {
-  name: string;
   email: string;
   password: string;
 };
 
-const RegisterPage = () => {
+export default function LoginPage({ }) {
   const router = useRouter();
-  const { registerUser } = useContext(AuthContext);
+  const { loginUser } = useContext(AuthContext);
 
   const {
     register,
@@ -26,15 +27,14 @@ const RegisterPage = () => {
     formState: { errors },
   } = useForm<FormData>();
   const [showError, setShowError] = useState(false);
-  const [errorMessage, setErrorMessage] = useState('');
 
-  const onRegisterForm = async ({ name, email, password }: FormData) => {
+  const onLoginUser = async ({ email, password }: FormData) => {
     setShowError(false);
-    const { hasError, message } = await registerUser(name, email, password);
 
-    if (hasError) {
+    const isValidLogin = await loginUser(email, password);
+
+    if (!isValidLogin) {
       setShowError(true);
-      setErrorMessage(message!);
       setTimeout(() => setShowError(false), 3000);
       return;
     }
@@ -46,7 +46,7 @@ const RegisterPage = () => {
   return (
     <AuthLayout title={'Ingresar'}>
       <form
-        onSubmit={handleSubmit(onRegisterForm)}
+        onSubmit={handleSubmit(onLoginUser)}
         noValidate
       >
         <Box sx={{ width: 350, padding: '10px 20px' }}>
@@ -62,7 +62,7 @@ const RegisterPage = () => {
                 variant='h1'
                 component='h1'
               >
-                Crear cuenta
+                Iniciar Sesión
               </Typography>
               <Chip
                 label='No reconocemos ese usuario / contraseña'
@@ -73,22 +73,6 @@ const RegisterPage = () => {
               />
             </Grid>
 
-            <Grid
-              item
-              xs={12}
-            >
-              <TextField
-                label='Nombre completo'
-                variant='filled'
-                fullWidth
-                {...register('name', {
-                  required: 'Este campo es requerido',
-                  minLength: { value: 2, message: 'Mínimo 2 caracteres' },
-                })}
-                error={!!errors.name}
-                helperText={errors.name?.message}
-              />
-            </Grid>
             <Grid
               item
               xs={12}
@@ -118,8 +102,14 @@ const RegisterPage = () => {
                 {...register('password', {
                   required: 'Este campo es requerido',
                   validate: isPassword,
-                  minLength: { value: 6, message: 'Mínimo 6 caracteres' },
-                  maxLength: { value: 10, message: 'Máximo 10 caracteres' },
+                  minLength: {
+                    value: 6,
+                    message: 'Debe tener al menos 6 caracteres',
+                  },
+                  maxLength: {
+                    value: 10,
+                    message: 'Debe tener menos de 10 caracteres',
+                  },
                 })}
                 error={!!errors.password}
                 helperText={errors.password?.message}
@@ -135,10 +125,10 @@ const RegisterPage = () => {
                 color='secondary'
                 className='circular-btn'
                 size='large'
-                fullWidth
                 style={{ backgroundColor: 'black' }}
+                fullWidth
               >
-                Registrar
+                Ingresar
               </Button>
             </Grid>
 
@@ -149,10 +139,10 @@ const RegisterPage = () => {
               justifyContent='end'
             >
               <NextLink
-                href='/auth/login'
+                href='/auth/register'
                 passHref
               >
-                ¿Ya tienes cuenta?
+                ¿No tienes cuenta?
               </NextLink>
             </Grid>
           </Grid>
@@ -162,4 +152,4 @@ const RegisterPage = () => {
   );
 };
 
-export default RegisterPage;
+
