@@ -1,23 +1,26 @@
+
+'use client';
 import NextLink from 'next/link';
-import { useRouter } from 'next/router';
 import { useContext, useState } from 'react';
 import { useForm } from 'react-hook-form';
 
-import { AuthLayout } from '@/components/layouts/AuthLayout';
-import { isEmail, isPassword } from '@/utils/validations';
 import { ErrorOutline } from '@mui/icons-material';
 import { Box, Button, Chip, Grid, TextField, Typography } from '@mui/material';
-
 import { AuthContext } from '../../context';
+import { AuthLayout } from '../../components/layouts/AuthLayout';
+import { isEmail, isPassword } from '../../utils/validations';
+import { useRouter } from 'next/navigation';
+
 
 type FormData = {
+  name: string;
   email: string;
   password: string;
 };
 
-const LoginPage = () => {
+export default function RegisterPage({ }) {
   const router = useRouter();
-  const { loginUser } = useContext(AuthContext);
+  const { registerUser } = useContext(AuthContext);
 
   const {
     register,
@@ -25,14 +28,15 @@ const LoginPage = () => {
     formState: { errors },
   } = useForm<FormData>();
   const [showError, setShowError] = useState(false);
+  const [errorMessage, setErrorMessage] = useState('');
 
-  const onLoginUser = async ({ email, password }: FormData) => {
+  const onRegisterForm = async ({ name, email, password }: FormData) => {
     setShowError(false);
+    const { hasError, message } = await registerUser(name, email, password);
 
-    const isValidLogin = await loginUser(email, password);
-
-    if (!isValidLogin) {
+    if (hasError) {
       setShowError(true);
+      setErrorMessage(message!);
       setTimeout(() => setShowError(false), 3000);
       return;
     }
@@ -44,7 +48,7 @@ const LoginPage = () => {
   return (
     <AuthLayout title={'Ingresar'}>
       <form
-        onSubmit={handleSubmit(onLoginUser)}
+        onSubmit={handleSubmit(onRegisterForm)}
         noValidate
       >
         <Box sx={{ width: 350, padding: '10px 20px' }}>
@@ -60,7 +64,7 @@ const LoginPage = () => {
                 variant='h1'
                 component='h1'
               >
-                Iniciar Sesión
+                Crear cuenta
               </Typography>
               <Chip
                 label='No reconocemos ese usuario / contraseña'
@@ -71,6 +75,22 @@ const LoginPage = () => {
               />
             </Grid>
 
+            <Grid
+              item
+              xs={12}
+            >
+              <TextField
+                label='Nombre completo'
+                variant='filled'
+                fullWidth
+                {...register('name', {
+                  required: 'Este campo es requerido',
+                  minLength: { value: 2, message: 'Mínimo 2 caracteres' },
+                })}
+                error={!!errors.name}
+                helperText={errors.name?.message}
+              />
+            </Grid>
             <Grid
               item
               xs={12}
@@ -100,14 +120,8 @@ const LoginPage = () => {
                 {...register('password', {
                   required: 'Este campo es requerido',
                   validate: isPassword,
-                  minLength: {
-                    value: 6,
-                    message: 'Debe tener al menos 6 caracteres',
-                  },
-                  maxLength: {
-                    value: 10,
-                    message: 'Debe tener menos de 10 caracteres',
-                  },
+                  minLength: { value: 6, message: 'Mínimo 6 caracteres' },
+                  maxLength: { value: 10, message: 'Máximo 10 caracteres' },
                 })}
                 error={!!errors.password}
                 helperText={errors.password?.message}
@@ -123,10 +137,10 @@ const LoginPage = () => {
                 color='secondary'
                 className='circular-btn'
                 size='large'
-                style={{ backgroundColor: 'black' }}
                 fullWidth
+                style={{ backgroundColor: 'black' }}
               >
-                Ingresar
+                Registrar
               </Button>
             </Grid>
 
@@ -137,10 +151,10 @@ const LoginPage = () => {
               justifyContent='end'
             >
               <NextLink
-                href='/auth/register'
+                href='/auth/login'
                 passHref
               >
-                ¿No tienes cuenta?
+                ¿Ya tienes cuenta?
               </NextLink>
             </Grid>
           </Grid>
@@ -149,5 +163,3 @@ const LoginPage = () => {
     </AuthLayout>
   );
 };
-
-export default LoginPage;
